whatpm/t/swml/forms-generic-1.dat

#data
[[#form:'':'':'']]
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <sw form>
|       input=""
|       option=""
|       template=""

#data
a[[#form:'':'':'']]
#errors
#document
| <html>
|   xmlns xmlns="http://www.w3.org/1999/xhtml"
|   <head>
|   <body>
|     <p>
|       "a"
|       <sw form>
|         input=""
|         option=""
|         template=""
